requirements.txt
together
pandas
python-docx
openpyxl
pdf2image
PyMuPDF
pytesseract
matplotlib
seaborn
filetype
streamlit
easyocr